Move "unit" data classes to CountryInfo file
