+++ app/src/main/java/de/westnordost/streetcomplete/data/meta/CountryInfo.kt
@@ -1,10 +1,31 @@
 package de.westnordost.streetcomplete.data.meta
 
-import de.westnordost.streetcomplete.osm.LengthUnit
-import de.westnordost.streetcomplete.quests.max_speed.SpeedMeasurementUnit
-import de.westnordost.streetcomplete.quests.max_weight.WeightMeasurementUnit
+import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import java.util.Locale
+
+@Serializable
+enum class LengthUnit(private val abbr: String) {
+    @SerialName("meter") METER("m"),
+    @SerialName("foot and inch") FOOT_AND_INCH("ft / in");
+
+    override fun toString() = abbr
+}
+
+@Serializable
+enum class SpeedMeasurementUnit(private val displayString: String) {
+    @SerialName("kilometers per hour") KILOMETERS_PER_HOUR("km/h"),
+    @SerialName("miles per hour") MILES_PER_HOUR("mph");
+
+    override fun toString() = displayString
+}
+
+@Serializable
+enum class WeightMeasurementUnit(val displayString: String) {
+    @SerialName("ton") TON("TONS"),
+    @SerialName("short ton") SHORT_TON("TONS"),
+    @SerialName("pound") POUND("POUNDS"),
+}
 
 @Serializable
 data class IncompleteCountryInfo(
