+++ app/src/main/java/de/westnordost/streetcomplete/measure/TakeMeasurementLauncher.kt
@@ -4,13 +4,13 @@ import android.app.Activity
 import android.content.Context
 import androidx.activity.result.ActivityResultCaller
 import androidx.activity.result.contract.ActivityResultContracts
+import de.westnordost.streetcomplete.data.meta.LengthUnit
 import de.westnordost.streetcomplete.measure.MeasureActivity.Companion.RESULT_MEASURE_FEET
 import de.westnordost.streetcomplete.measure.MeasureActivity.Companion.RESULT_MEASURE_INCHES
 import de.westnordost.streetcomplete.measure.MeasureActivity.Companion.RESULT_MEASURE_METERS
 import de.westnordost.streetcomplete.osm.Length
 import de.westnordost.streetcomplete.osm.LengthInFeetAndInches
 import de.westnordost.streetcomplete.osm.LengthInMeters
-import de.westnordost.streetcomplete.osm.LengthUnit
 import de.westnordost.streetcomplete.util.ActivityForResultLauncher
 
 /** Launches the MeasureActivity and returns its result */
