--- app/src/main/java/de/westnordost/streetcomplete/quests/max_speed/SpeedMeasurementUnit.kt
@@ -1,12 +1,0 @@
-package de.westnordost.streetcomplete.quests.max_speed
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-enum class SpeedMeasurementUnit(private val displayString: String) {
-    @SerialName("kilometers per hour") KILOMETERS_PER_HOUR("km/h"),
-    @SerialName("miles per hour") MILES_PER_HOUR("mph");
-
-    override fun toString() = displayString
-}
